requirements.txt
Flask
Flask-CORS
google-cloud-speech
google-cloud-translate
google-cloud-storage
